README.md: Remove host and port specification from flask run command
For running locally, there is no need to bind to every network interface
and make the app visible on your local network, and we shouldn't
recommend people do this unless they understand what they are doing.

Also, switch back to the standard flask port 5000; I don't know a reason
to change it.