Downgraded my competencies!
I originally filled out *interests* as "interests". I missed the how
to misuse the term as specified in the Readme file.

Recently, having fielded a bunch of requests for help for a C++ based
project (I last used C++ in a project aeons[>=20yrs] ago) and which
turned out to be an HPC optimization problem and then another which
turned out to be a visualization/GUI based problem, I felt that I need
to tone-down my skills graph.

I've also raised an issue in this repo to enritch the relation
vocabulary as it's at best confusing at worst not rich enough for the
purpose intended: there are valid reasons to be interested in a tech
as opposed to being proficient.