application.py: Remove explicit error handling
In the old code, any exception was caught and replaced with a print
statement followed by exiting the program. For a web app, we probably
don't want to exit the program. Instead, we will let the exception be
handled by the web app runner. This allows the flask debugger, for example,
to be used.