Dockerfile: Improve healthcheck
We use newer features in Docker to initially try the healthcheck with a
short interval, but then reduce to a longer interval for normal use.
This allows us to have the service marked as healthy as soon as it is,
but not to spam the index page too often (which results in an API call
to CapX).